+++ app/src/main/java/com/techcourse/AppWebApplicationInitializer.java
@@ -3,7 +3,7 @@ package com.techcourse;
 import jakarta.servlet.ServletContext;
 import java.util.List;
 import nextstep.mvc.DispatcherServlet;
-import nextstep.mvc.controller.asis.ManualHandlerAdapter;
+import nextstep.mvc.controller.tobe.ManualHandlerAdapter;
 import nextstep.mvc.controller.tobe.AnnotationHandlerAdapter;
 import nextstep.mvc.controller.tobe.AnnotationHandlerMapping;
 import nextstep.web.WebApplicationInitializer;
