+++ mvc/src/main/java/nextstep/mvc/controller/tobe/ManualHandlerAdapter.java
@@ -1,8 +1,9 @@
-package nextstep.mvc.controller.asis;
+package nextstep.mvc.controller.tobe;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import nextstep.mvc.HandlerAdapter;
+import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
